hardware: add joint I/O, torque/power APIs, diagnostics; fix warnings
livelybot_serial:
- Publish standard joint_states (stale motors dropped, not NaN)
- Subscribe joint_command (generic, replaces DynamixelController/command)
- Subscribe set_torque_individual (JointTorque) for teach/puppeteering mode
- Add core/switch_power SetBool service and core/power_switch_status topic
  matching the HCM motor-power cut API
- Add per-motor ROS diagnostics: connection STALE, fault ERROR,
  torque-overload WARN; drop unsupported temperature fields
- Rename JointCommand.max_currents → max_torques; update all callers
- Fix all compiler warnings: sign-compare, reorder, missing-initializer,
  unused variable/parameter

+++ src/lib/livelybot_hardware_sdk/src/livelybot_serial/src/hardware/canboard.cc
@@ -13,7 +13,7 @@ canboard::canboard(int _CANboard_ID, std::vector<lively_serial *> *ser, rclcpp::
         RCLCPP_ERROR(node_->get_logger(), "Failed to get params CANport_num");
     }
 
-    for (size_t j = 1; j <= CANport_num; j++)  // one serial port per CAN port
+    for (int j = 1; j <= CANport_num; j++)  // one serial port per CAN port
     {
         CANport.push_back(new canport(j, _CANboard_ID,
                                      (*ser)[(_CANboard_ID - 1) * CANport_num + j - 1], node_));
